Fix a couple of warnings
git-svn-id: http://svn.python.org/projects/python/branches/ssize_t@41786 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Modules/_bsddb.c
@@ -1097,7 +1097,7 @@ _db_associateCallback(DB* db, const DBT* priKey, const DBT* priData,
         }
         else if (PyString_Check(result)) {
             char* data;
-            int   size;
+            Py_ssize_t size;
 
             CLEAR_DBT(*secKey);
 #if PYTHON_API_VERSION <= 1007
